server/opinion/api-2/requirements.txt
PyJWT
PyMuPDF
azure-functions
azure-data-tables
azure-storage-blob
requests
tiktoken
langchain
langchain-text-splitters
langchain-community
requests-toolbelt
azure-ai-documentintelligence
sqlalchemy
psycopg2-binary
python-docx
docxtpl
azure-communication-email
beautifulsoup4
numpy